Design mockup, add some method stubs and timer
- Added a minimal visual mockup of some colored buttons and a "start"
button at the first page, and a empty "about"-page.
- Added "mode" variable, for indicating the state in which the program
is. Suggestion for possible states: mode = "menu": before or after a
game, "preview": the colors are shown, "waiting": the app waits for user
to click on the right colors.

- At the moment, pressing the "start"-button triggers a timer and counts
seconds. Could be used for switching states and previewing the buttons
which the user has to press.